core: Fix host_init_verifier for devices with overriding services
Previous commit e5f0520 did not fully solve this, the build can still
fail if the non-overriden definition gets parsed _after_ the overriden
one.

Also change an error return to just a log warning, was running into
services failing out at this part when it should just ignore it and
move on.

Test: m dist
Change-Id: I038e2aab67ef627c5865c78337cee9a268c78255
Signed-off-by: HDzungx <hdzungx@gmail.com>